Allow online installation of packages (#56)
## Registries

_See the newly added registry.nuon file for more specific info about the
registry format._

This PR adds a new concept: **registry**. Registries are .nuon files
containing a list of packages, their versions, and how they should be
installed. Currently, registry can contain two kinds of packages:
* `local` -- installed from local file system
* `git` -- installed from remote Git repository

The list of recognized registries is stored in `$env.NUPM_REGISTRIES` as
a record of name: path pairs. That way, users can maintain their
collection of sources from where they want to install packages. Not sure
if this will be the final way to pass available registries to nupm, but
should be good enough for early testing.

## Online Installation

If a package is to be installed from a Git repository, the repository is
cached in `$env.NUPM_CACHE` to avoid re-cloning the same repository all
the time. Once cloned, the local path of the cloned repository is passed
to the installer, as if you called `nupm install --path=...`.

One important missing feature is some kind of file integrity check. One
step at a time...

## Versions

If a registry contains multiple versions (common scenario), by default,
the newest one is installed. "Newest" currently means sorting the
versions alphabetically and taking the last one. If `--pkg-version` is
specified, nupm tries to install the exact version passed to this
option. This needs better version matching (e.g., version 0.2 matching
0.2.1 etc.), but again, one step at a time...

## New command

`nupm search` searches for packages in registries without installing
them. I thought this would be useful to have early on.

## Try it out

From the repository root:

```nushell
overlay use nupm/
nupm install nu-git-manager
# use nu-git-manager *
# ...
```

## TODO

- [x] Remove debug stuff
- [x] Nicer / more informative prints
- [x] (can be done over time) Rewrite
https://github.com/nushell/nupm/blob/main/index.nuon to fit the registry
format, move it to its own repo
	- It would be a single-file repository
- Package authors would put PRs there whenever they update their package
- [x] Tests

## After
- better version matching => `0.2` shouldn't match `0.2.1`
- file integrity checks