Import std/log environment; CI: Update Ubuntu (#103)
Will fix [this issue](https://github.com/nushell/nushell/issues/14109)
but its not a good fix anyway

## Description

This is a hotfix PR to address an unusual issue with the `std log`
package in Nushell. When the log is imported during the initialization
stage, it doesn’t persist long enough to be fully available at runtime,
specifically, the environment variables seem to disappear, although the
log function itself remains accessible. This is a temporary fix and
might not be the final solution. It could potentially be a bug in
Nushell, but unfortunately, I don't have the time to investigate it
further.

for more information please check [this
issue](https://github.com/nushell/nushell/issues/14109)

I think this fix changes the output of the nupm, Plz close this PR if
you found any real solution