Fix: Hash detection affects all Search results (#107)
## Description

The current Hash check in the code will directly cause the program to
throw an error. When using Search, if there is even one Hash check
failure, no information can be displayed.

I have fixed this issue and retained the error throw in the Install
section. However, I don't think it's necessary to make the program fail
to run specifically because of this issue.

The possible useful solutions I have thought of, which I can modify if
you are willing to accept, are:

1. Automatically ignore all packages with Hash check failures.
2. Give users the right to choose, informing them that the content does
not match, but they can still install it.